LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/worker-src-wildcard/worklet-audio-import-data.https-expected.txt

CONSOLE MESSAGE: Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=GENERATED_KEY&path=/mixed-content
CONSOLE MESSAGE: Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=GENERATED_KEY&path=/mixed-content
CONSOLE MESSAGE: Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=GENERATED_KEY&path=/mixed-content
CONSOLE MESSAGE: Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?action=purge&key=GENERATED_KEY&path=%2Fmixed-content

FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Importing a module script is canceled."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and swap-origin redirection from https context.: securitypolicyviolation
